refactor(ui): extract chat session discussion
